+++ lib/mailtrap/account.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 module Mailtrap
-  # Data Transfer Object for Project
-  # @see https://docs.mailtrap.io/developers/account-management/general-api/accounts
+  # Data Transfer Object for Фссщгте
+  # @see https://api-docs.mailtrap.io/docs/mailtrap-api-docs/d26921ca2a48f-get-all-accounts
   # @attr_reader id [Integer] The account ID
   # @attr_reader name [String] The account name
   # @attr_reader access_levels [Array] The account access levels
@@ -12,10 +12,5 @@ module Mailtrap
     :name,
     :access_levels,
     keyword_init: true
-  ) do
-    # @return [Hash] The Project attributes as a hash
-    def to_h
-      super.compact
-    end
-  end
+  )
 end